--- src/CRUMBS.h
@@ -1,8 +1,0 @@
-#pragma once
-#include "crumbs/crumbs.h"
-#include "crumbs/crumbs_i2c.h"
-#include "crumbs/crumbs_crc.h"
-#include "crumbs/crumbs_message.h"
-#ifdef ARDUINO
-#include "crumbs/crumbs_arduino.h"
-#endif
