+++ src/crumbs_core.h
